requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==4.2.13
django-allauth==0.63.3
django-environ==0.11.2
django-rest-knox==4.2.0
djangorestframework==3.15.2
drf-yasg==1.21.7
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
httplib2==0.22.0
idna==3.7
inflection==0.5.1
msgpack==1.0.8
oauthlib==3.2.2
packaging==24.1
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2